Bug fix: Store HBO ID as primary key in DB, as expected by HboApi.
HboApi expects episodes to have a column corresponding to the HBO ID,
but that was not stored in the DB at all. Instead we had an auto-
incrementing primary key _ID. The new schema uses a distinct column name
just to make it super clear it's an HBO ID, not some random integer.