Move all initialization out of Application class, it's not safe anymore.
Application#onCreate is called for every process in the app -- e.g., the
SyncAdapter process -- so access to any out-of-process state (db, shared
prefs, filesystem) is risky.

The more reliable way is to initialize stuff in ContentProvider#onCreate
because that is only called once, during app launch, and *only* from the
main thread.

This was prompted by the need to integrate Firebase Crash Reporting,
but is a good thing anyway.